nixos: added home-manager support
Maybe this is a little broken and non-nix but this seems to work for me
so far. Add home-manager configuration in .config/home.nix but have it
still managed by /etc/nixos/configuration.nix.

This is still not perfect since these two files will need to be placed
in their correct location to get things working.